corrected a bug in multistat, didn't generate non-covariate comparisons properly. Integrated mcsa to multistat.

+++ a5runCombiningRuns.sh
@@ -2,7 +2,7 @@
 
 if [ $# -ne 3 ]; then echo "Usage: $0 <scriptLocation> <text with list of matlab commands> <outputFile>"; exit; fi
 
-rm "$3"
+if  [ -f "$3" ]; then rm "$3"; fi
 for i in `cat $2`; do
 	outputFile=`echo $i | sed "s:combiningRuns('::" | sed "s:',':data2/subject:" | sed "s:',':_session:" | sed "s:');:_multi_t.mnc:"`
 	echo "${outputFile}:$i" >> $3
